requirements-fixed.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-PyMongo==3.0.1
Flask-WTF==1.2.2
Flask-Limiter==3.12
python-dotenv==1.0.1
Werkzeug==3.0.2
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
email-validator==2.1.0
psycopg2-binary==2.9.9
pymongo==4.6.2
flask-socketio==5.3.5
